move more methods and rename
